% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guyot.R
\name{guyot}
\alias{guyot}
\title{Algorithm of Guyot et al from http://www.ncbi.nlm.nih.gov/pmc/articles/PMC3313891/}
\usage{
guyot(digizeit, pub.risk, tot.events = "NA", tags = list())
}
\description{
Algorithm of Guyot et al from http://www.ncbi.nlm.nih.gov/pmc/articles/PMC3313891/
}

